Revert "Typrescript"
