+++ cypress/support/e2e.js
@@ -15,7 +15,7 @@
 
 // Import commands.js using ES2015 syntax:
 import './commands'
-/// <reference path="./index.d.ts" />
 
-import { csv2html } from '../../src'
+const { csv2html } = require('../../src')
 csv2html()
+
